requirements.txt
# -----------------------------
# Web / API
# -----------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# -----------------------------
# gRPC / Protobuf
# -----------------------------
grpcio>=1.60,<2.0
grpcio-tools>=1.60,<2.0
protobuf>=4.25,<6.0

# -----------------------------
# Data validation
# -----------------------------
pydantic>=2.6,<3.0

# -----------------------------
# Async / Networking
# -----------------------------
httpx>=0.27,<1.0

# -----------------------------
# Observability (minimal)
# -----------------------------
structlog>=24.1,<25.0

# -----------------------------
# Utilities
# -----------------------------
python-dotenv>=1.0,<2.0
